[noah.u] 인증 관리 Step1, 2 (#8)
* feat: AuthController 구현 & 롬복 의존성 추가

* feat: AuthService 토큰 생성 메서드 구현

* feat: MemberController 구현

* feat: MemberService 구현

* test: Member 조회 테스트 작성

* feat: LoginInterceptor & LoginArgumentResolver 작성

* feat: 커스텀 예외 정의

* test: AuthE2ETest 요구사항에 맞게 수정

* feat: 멤버 로그인시 예외처리 구현

* test: 테스트 api 요청시 토큰 헤더에 추가하도록 수정

* test: Member 조회 테스트 작성

* feat: 자신이 예약이 아닌 경우 예약 취소가 불가능한 테스트 추가 및 구현

* style: 패키지 분리

* feat: AuthorizationExtractor 적용

* feat: 인증 방식을 name에서 id로 변경

* feat: 리팩토링 + 바뀐 인증 방식에서 동작하도록 변경

* docs : README 추가

* feat: Bearer를 attribute에 set할 수 있도록 변경

* refactor: @Value 어노테이션 적용

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,13 +1,18 @@
 package nextstep.reservation;
 
+import nextstep.member.Member;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
-import nextstep.support.DuplicateEntityException;
+import nextstep.support.exception.AuthorizationException;
+import nextstep.support.exception.DuplicateReservationException;
+import nextstep.support.exception.NoSuchReservationException;
+import nextstep.support.exception.NotReservationOwnerException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class ReservationService {
@@ -21,7 +26,7 @@ public class ReservationService {
         this.scheduleDao = scheduleDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
+    public Long create(ReservationRequest reservationRequest, Long memberId) {
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
         if (schedule == null) {
             throw new NullPointerException();
@@ -29,12 +34,13 @@ public class ReservationService {
 
         List<Reservation> reservation = reservationDao.findByScheduleId(schedule.getId());
         if (!reservation.isEmpty()) {
-            throw new DuplicateEntityException();
+            throw new DuplicateReservationException();
         }
 
         Reservation newReservation = new Reservation(
                 schedule,
-                reservationRequest.getName()
+                reservationRequest.getName(),
+                memberId
         );
 
         return reservationDao.save(newReservation);
@@ -49,10 +55,15 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Long memberId) {
         Reservation reservation = reservationDao.findById(id);
+
         if (reservation == null) {
-            throw new NullPointerException();
+            throw new NoSuchReservationException();
+        }
+
+        if(!Objects.equals(reservation.getMemberId(), memberId)) {
+            throw new NotReservationOwnerException();
         }
 
         reservationDao.deleteById(id);
